lib creation
